+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+Version 1.x.y - 2021-MM-DD
+==========================
+
+- Added cost estimation to all adapters
+
 Version 1.0.0 - 2021-08-18
 ==========================